api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pydantic-settings==2.6.1
Unidecode==1.3.8
requests==2.32.3
pypdf==5.1.0
itsdangerous==2.2.0
pymssql==2.3.2
